+++ internal/component/prometheus/exporter/windows/config_stubs.go
@@ -1,0 +1,6 @@
+//go:build !windows
+// +build !windows
+
+package windows
+
+var defaultTextFileDirectories = []string{}